fix: Always skip exactly one event
The previous implementation was using a numeric value of s:skip which
specified how many events to skip. Hence, the variable was *incremented*
by each mapping, even if the cursor hadn't move.

This meant that using a search command when there was only one match and
the cursor it onto it, or when there was't even a match caused the
'hlsearch' to go away after more than one cursor move, which was
unpleasant.

Now however, the s:skip variable is simpler: it is merely a boolean
value and can thus only skip *one* event. The only case which would
result in an event being skipped when it shouldn't are search that don't
move the cursor (n when there's only one match, or search for one match
when the cursor is already onto it).

This still isn't perfect though, and I might completely change the
strategy of vim-chimera